FE-845 Wrap gulp command line with "drcp"
