+++ wfh/lib/gulp/git-hooks/pre-commit
