+++ wfh/lib/gulp/templates/app-template.js
