+++ wfh/lib/gulp/templates/config.local-template.yaml
