+++ wfh/lib/gulp/templates/gulpfile-template.js
