+++ wfh/lib/gulp/templates/log4js.json
