+++ wfh/lib/gulp/templates/module-resolve.browser.tmpl.js
