+++ wfh/lib/gulp/templates/module-resolve.server.tmpl.js
