+++ wfh/lib/gulp/templates/package.json.template
