+++ wfh/package.json
@@ -1,6 +1,6 @@
 {
   "name": "dr-comp-package",
-  "version": "0.1.0-alpha.5",
+  "version": "0.1.0-alpha.7",
   "description": "DR web component package developer toolkit",
   "main": "lib/main.js",
   "bin": "bin/dr.js",
